+++ modules/040-control-plane-manager/hooks/audit_policy_basic_targets_generated.go
@@ -131,8 +131,8 @@ var auditPolicyBasicServiceAccounts = []string{
 	"system:serviceaccount:d8-system:terraform-state-exporter",
 	"system:serviceaccount:d8-system:webhook-handler",
 	"system:serviceaccount:d8-user-authn:basic-auth-proxy",
+	"system:serviceaccount:d8-user-authn:controller",
 	"system:serviceaccount:d8-user-authn:dex",
-	"system:serviceaccount:d8-user-authn:user-authn-controller",
 	"system:serviceaccount:d8-user-authz:permission-browser-apiserver",
 	"system:serviceaccount:d8-user-authz:webhook",
 	"system:serviceaccount:kube-system:d8-control-plane-manager",
