--- modules/150-user-authn/images/user-authn-controller/src/cmd/health.go
@@ -1,51 +1,0 @@
-/*
-Copyright 2026 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-	"time"
-
-	"sigs.k8s.io/controller-runtime/pkg/cache"
-	"sigs.k8s.io/controller-runtime/pkg/healthz"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-)
-
-const cacheSyncCheckTimeout = 2 * time.Second
-
-func addHealthChecks(mgr manager.Manager) error {
-	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
-		return fmt.Errorf("add healthz check: %w", err)
-	}
-	if err := mgr.AddReadyzCheck("cache-sync", cacheSyncCheck(mgr.GetCache())); err != nil {
-		return fmt.Errorf("add readyz check: %w", err)
-	}
-	return nil
-}
-
-func cacheSyncCheck(c cache.Cache) healthz.Checker {
-	return func(req *http.Request) error {
-		ctx, cancel := context.WithTimeout(req.Context(), cacheSyncCheckTimeout)
-		defer cancel()
-		if !c.WaitForCacheSync(ctx) {
-			return fmt.Errorf("informer cache is not synced")
-		}
-		return nil
-	}
-}
